refactor: move RunLogger to src/core for #29
- Create src/core/run_logger.py as the canonical location
- Update src/bugfix_agent/run_logger.py to re-export for backward compatibility
- Export RunLogger from src/core/__init__.py
- Add import path tests to verify all import patterns work
